# docs/swagger.yaml
...
swagger: '2.0'
info:
  version: '1.0.0'
  title: Factory Management API - Employees
  description: API documentation for employee-related operations in the Factory Management System.
paths:
...
  /analytics/employee-performance:
    get:
      summary: Analyze employee performance
      description: Analyze the total quantity of products each employee has produced.
      responses:
        200:
          description: Success
        500:
          description: Internal server error
  /analytics/top-products:
    get:
      summary: Get top-selling products
      description: Fetch top-selling products based on total quantity ordered.
      responses:
        200:
          description: Success
        500:
          description: Internal server error
  /analytics/customer-lifetime-value:
    get:
      summary: Calculate customer lifetime value
      description: Calculate the total value of orders placed by each customer.
      parameters:
        - name: threshold
          in: query
          type: number
          description: Minimum total order value to filter customers (default:1000)
      responses:
        200:
          description: Success
        400:
          description: Invalid threshold value
        500:
          description: Internal server error
  /analytics/production-efficiency:
    get:
      summary: Evaluate production efficiency
      description: Evaluate production efficiency by calculating the total quantity produced for each product on a specific date.
      parameters:
        - name: date
          in: query
          type: string
          description: Date in YYYY-MM-DD format (required)
      responses:
        200:
          description: Success
        400:
          description: Date is required
        500:
          description: Internal server error
  /customers:
    post:
      summary: Create a new customer
      description: Create a customer with the specified details.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - name
              - email
              - phone
            properties:
              name:
                type: string
                description: Customer's name.
              email:
                type: string
                description: Customer's email.
              phone:
                type: string
                description: Customer's phone number.
      responses:
        201:
          description: Customer created successfully.
        400:
          description: Validation error.
        500:
          description: Internal server error.
    get:
      summary: Get paginated customers
      description: Retrieves paginated customers with optional sorting and metadata.
      parameters:
        - name: page
          in: query
          type: integer
          description: Page number (default:1).
        - name: per_page
          in: query
          type: integer
          description: Items per page (default:10, max:100).
        - name: sort_by
          in: query
          type: string
          description: Field to sort by ('name','email','phone') (default:'name').
        - name: sort_order
          in: query
          type: string
          description: Sort order ('asc', 'desc') (default:'asc').
        - name: include_meta
          in: query
          type: boolean
          description: Include metadata (default:true).
      responses:
        200:
          description: Paginated list of customers retrieved successfully.
        400:
          description: Validation error.
        500:
          description: Internal server error.
  /customers/{customer_id}:
    get:
      summary: Get customer by ID
      description: Retrieve a customer by their ID.
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: ID of the customer to retrieve.
      responses:
        200:
          description: Customer retrieved successfully.
        404:
          description: Customer not found.
    put:
      summary: Update customer by ID
      description: Update a customer's details by their ID.
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: ID of the customer to update.
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: Updated name of the customer.
              email:
                type: string
                description: Updated email of the customer.
              phone:
                type: string
                description: Updated phone number of the customer.
      responses:
        200:
          description: Customer updated successfully.
        400:
          description: Validation error.
        404:
          description: Customer not found.
    delete:
      summary: Delete customer by ID
      description: Delete a customer by their ID.
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: ID of the customer to delete.
      responses:
        200:
          description: Customer deleted successfully.
        404:
          description: Customer not found.
...